+++ src/Recyclarr.Core/Settings/Models/RecyclarrSettings.cs
@@ -3,8 +3,6 @@ namespace Recyclarr.Settings.Models;
 [UsedImplicitly(ImplicitUseKindFlags.Assign, ImplicitUseTargetFlags.WithMembers)]
 public record RecyclarrSettings
 {
-    // Replaced by ResourceProviders
-    public Repositories? Repositories { get; init; }
     public IReadOnlyCollection<ResourceProvider> ResourceProviders { get; init; } = [];
     public bool EnableSslCertificateValidation { get; init; } = true;
     public LogJanitorSettings LogJanitor { get; init; } = new();